Added BookDetail container with activeBook reducer
So, whenever a new item in bookList container is clicked, action creator selectBook is called which throws an object with type 'BOOK_SELECTED'. This object is received by each reducer ( in this case reducer_active_book ) which can then determine if the reducer has any use for that object. If it does, it will accept it and return the appropriate  value for the application state's property back to the index.js thereby effectively setting the application state which results in BookDetail re-rendering itself as React tells BookDetail that the application state's property (activeBook) has changed.