Added BookDetail container with activeBook reducer
So, whenever a new item in bookList container is clicked, action creator selectBook is called which throws an object with type 'BOOK_SELECTED'. This object is received by each reducer ( in this case reducer_active_book ) which can then determine if the reducer has any use for that object. If it does, it will accept it and send the appropriate application's state property back to the container ( in this case, bookDetail ) where it is required.